requirements.txt
streamlit
pandas
yfinance
matplotlib
feedparser
gspread
google-auth
plotly